refactor: bounty endpoint.
