+++ src/bounty/bounty.controller.ts
@@ -1,28 +1,26 @@
 import { Body, Controller, Post } from '@nestjs/common';
 import { ApiBody } from '@nestjs/swagger';
-import { SubstrateService } from 'src/substrate/substrate.service';
-import { BountyService } from './bounty.service';
-import { CreateBountyDto } from './dto/create-bounty.dto';
+import { Repository } from 'typeorm';
+import { InjectRepository } from '@nestjs/typeorm';
+import { DataStakingDto } from './dto/data-staking.dto';
+import { DataStakingEvents } from './models/data-staking-events.entity';
 
 @Controller('bounty')
 export class BountyController {
   constructor(
-    private substrateService: SubstrateService,
-    private readonly bountyService: BountyService,
+    @InjectRepository(DataStakingEvents)
+    private readonly dataStakingEventsRepository: Repository<DataStakingEvents>,
   ) {}
 
-  @Post('/create-data-bounty')
-  @ApiBody({ type: CreateBountyDto })
-  async create(@Body() data: CreateBountyDto) {
-    const resultDataBounty = await this.bountyService.create(data);
-
-    await this.substrateService.submitStaking(
-      resultDataBounty.hash_bounty_ocean,
-      data.order_id,
-    );
-
-    return {
-      data: resultDataBounty,
-    };
+  @Post('/create-sync-event')
+  @ApiBody({ type: DataStakingDto })
+  async CreateSyncEvent(@Body() dto: DataStakingDto) {
+    const dataStakingEvents = new DataStakingEvents();
+    dataStakingEvents.order_id = dto.order_id;
+    dataStakingEvents.service_category_id = dto.service_category_id;
+    dataStakingEvents.filename = dto.filename;
+    dataStakingEvents.created_at = new Date();
+    dataStakingEvents.event_processed = false;
+    return this.dataStakingEventsRepository.save(dataStakingEvents);
   }
 }
