+++ src/bounty/bounty.module.ts
@@ -1,24 +1,12 @@
-import { forwardRef, Module } from '@nestjs/common';
-import { BountyService } from './bounty.service';
+import { Module } from '@nestjs/common';
 import { BountyController } from './bounty.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { DataBounty } from './models/bounty.entity';
-import { ElasticsearchModule } from '@nestjs/elasticsearch';
-import { EscrowAccounts } from 'src/escrow/models/deposit.entity';
-import { SubstrateModule } from 'src/substrate/substrate.module';
+import { DataStakingEvents } from './models/data-staking-events.entity';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([DataBounty]),
-    TypeOrmModule.forFeature([EscrowAccounts]),
-    forwardRef(() => SubstrateModule),
-    ElasticsearchModule.registerAsync({
-      useFactory: async () => ({
-        node: process.env.ELASTICSEARCH_NODE,
-      }),
-    }),
+    TypeOrmModule.forFeature([DataStakingEvents]),
   ],
-  providers: [BountyService],
   controllers: [BountyController],
 })
 export class BountyModule {}
