--- src/bounty/dto/create-bounty.dto.ts
@@ -1,15 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-
-export class CreateBountyDto {
-  @ApiProperty({
-    type: String,
-    description: 'bounty_ocean',
-  })
-  bounty_ocean: string;
-
-  @ApiProperty({
-    type: String,
-    description: 'order_id',
-  })
-  order_id: string;
-}
