+++ src/cloud-storage/cloud-storage.controller.ts
@@ -1,31 +1,13 @@
 import { Controller, Get, Query, Post, Body } from '@nestjs/common';
 import { GCloudStorageService } from '@aginix/nestjs-gcloud-storage';
-import { Repository } from 'typeorm';
-import { ApiBody } from '@nestjs/swagger';
-import { InjectRepository } from '@nestjs/typeorm';
-import { DataStakingSyncEvents } from './models/data-staking-sync-events.entity';
-import { DataStakingDto } from './dto/data-staking.dto';
 
 @Controller('gcs')
 export class CloudStorageController {
   constructor(
-    @InjectRepository(DataStakingSyncEvents)
-    private readonly dataStakingSyncEventsRepository: Repository<DataStakingSyncEvents>,
     private readonly cloudStorageService: GCloudStorageService,
   ) {}
 
-  @Post('create_sync_event')
-  @ApiBody({ type: DataStakingDto })
-  async CreateSyncEvent(@Body() dto: DataStakingDto) {
-    //TODO: Discuss return data
-    const dataStakingSyncEvents = new DataStakingSyncEvents();
-    dataStakingSyncEvents.filename = dto.filename;
-    dataStakingSyncEvents.created_at = new Date();
-    dataStakingSyncEvents.event_processed = false;
-    return this.dataStakingSyncEventsRepository.save(dataStakingSyncEvents);
-  }
-
-  @Get('signed_url')
+  @Get('/signed-url')
   async GetSignedUrl(
     @Query('filename') filename: string,
     @Query('action') action: string,
