+++ src/cloud-storage/cloud-storage.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { GCloudStorageModule } from '@aginix/nestjs-gcloud-storage';
 import { CloudStorageController } from './cloud-storage.controller';
-import { DataStakingSyncEvents } from './models/data-staking-sync-events.entity';
 
 require('dotenv').config(); // eslint-disable-line
 
@@ -13,9 +12,7 @@ require('dotenv').config(); // eslint-disable-line
       storageBaseUri: process.env.STORAGE_BASE_URI,
       predefinedAcl: 'private',
     }),
-    TypeOrmModule.forFeature([DataStakingSyncEvents]),
   ],
-  exports: [TypeOrmModule],
   controllers: [CloudStorageController],
 })
 export class CloudStorageModule {}
