--- src/cloud-storage/dto/data-staking.dto.ts
@@ -1,9 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-
-export class DataStakingDto {
-  @ApiProperty({
-    type: String,
-    description: 'filename',
-  })
-  filename: string;
-}
